Framewrok/Report/behaveDemo/behaveDemo/requirements.txt
allure-behave==2.5.0
allure-python-commons==2.5.0
attrs==18.1.0
behave==1.2.6
parse==1.8.4
parse-type==0.4.2
pluggy==0.7.1
PyHamcrest==1.9.0
selenium==3.14.0
six==1.11.0
urllib3==1.23
